Split package and update to v0.2.1
The revised recipe splits up the package according to the subpackages of
the PyPI package. Currently only the subpackages shimmy-gym and
shimmy-atari are available. The tests required a patch to skip the
rendering, which is not available without additional work during build.

The other subpackages have been prepared, but their builds are skipped
mainly due to missing dependencies on conda-forge.

Overall the recipe is analog to the Gymnasium feedstock recipe.