Update recipe to fix broken package
This adds dependencies to the host section of shimmy. This avoids pip
having to install these, which caused the dependencies (like numpy and
gymnasium) to be packaged as well into shimmy.

gymnasium has been added as run dependency, since importing shimmy also
imports gymnasium and it is now not included in the package anymore.

For the test of shimmy-atari the python version restriction could be
lifted by updating the version of autorom. autorom v0.6.1 is not
dependend on libtorrent anymore, see
https://github.com/Farama-Foundation/AutoROM/releases/tag/v0.6.1
libtorrent is only available for Python <= 3.10 currently. Somehow the
restriction to Python <= 3.10 broke the test of shimmy-atari, because it
could not install shimmy from the build_artifacts.